+++ examples/example-antd-management-fast-framework/src/modelBuilders/formDesign.js
@@ -13,6 +13,7 @@ import {
   saveFormRemarkData,
   setData,
   setDataSchemaData,
+  setDocumentSchemaData,
 } from '../services/formDesign';
 
 export function buildModel() {
@@ -86,6 +87,32 @@ export function buildModel() {
         { call, put },
       ) {
         const response = yield call(setDataSchemaData, payload);
+
+        const dataAdjust = pretreatmentRemoteSingleData({
+          source: response,
+          successCallback: pretreatmentSuccessCallback || null,
+          failCallback: pretreatmentFailCallback || null,
+        });
+
+        yield put({
+          type: reducerNameCollection.reducerRemoteData,
+          payload: dataAdjust,
+          alias,
+          ...reducerDefaultParameters,
+        });
+
+        return dataAdjust;
+      },
+      *setDocumentSchema(
+        {
+          payload,
+          alias,
+          pretreatmentSuccessCallback,
+          pretreatmentFailCallback,
+        },
+        { call, put },
+      ) {
+        const response = yield call(setDocumentSchemaData, payload);
 
         const dataAdjust = pretreatmentRemoteSingleData({
           source: response,